Share AnimalSuperBuilderTest and repoint the docs at the TCK
The last forked suite test. The source suite tested the generated builder's
introspection, the ASM suite tested the generated bridge methods and
dispatching through the root builder's erasure; neither ran the other's
tests. Both sets pass on all three backends, so keep the union in the TCK.

The documentation includes source files by path, and moving files into
test-suite-tck broke fourteen of them, so repoint those includes. Two staged
builder snippets used the snippet:: macro, which only looks under
test-suite-<language> and so can no longer find the Java sources; spell out
the java and kotlin tabs it would have generated, keeping both languages on
the page. The docs build is clean apart from warnings that predate this
work: Elephant.java opens the tag `clazs` twice without closing it, and the
GenerateInterfaceBuilder snippet has only a Java source.

Every backend gains tests and none loses any: against the state before this
series the three suites gain fourteen tests between them and lose none.
