+++ src/main/docs/guide/annotations/builder.adoc
@@ -2,14 +2,14 @@ If you annotate a Java Record with `@Builder`, a `PersonBuilder` class is genera
 
 [source,java]
 ----
-include::test-suite-java/src/main/java/io/micronaut/sourcegen/example/Person.java[tags=clazz,indent=0]
+include::test-suite-tck/src/main/java/io/micronaut/sourcegen/example/Person.java[tags=clazz,indent=0]
 ----
 
 You can use a https://blogs.oracle.com/javamagazine/post/exploring-joshua-blochs-builder-design-pattern-in-java[builder pattern] to create a `Person` instance:
 
 [source,java]
 ----
-include::test-suite-java/src/test/java/io/micronaut/sourcegen/example/PersonBuilderTest.java[tags=test,indent=0]
+include::test-suite-tck/src/test/java/io/micronaut/sourcegen/example/PersonBuilderTest.java[tags=test,indent=0]
 ----
 
 NOTE: The https://github.com/micronaut-projects/micronaut-sourcegen/blob/master/sourcegen-generator/src/main/java/io/micronaut/sourcegen/generator/visitors/BuilderAnnotationVisitor.java[BuilderAnnotationVisitor] is an example of how to use the Micronaut SourceGen API.
